requirements.txt
python-telegram-bot==20.3
solana==0.30.0
solders==0.16.1
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.8.5